Backend completed(Tested)

+++ src/dto/ItemDTO.java
@@ -3,7 +3,7 @@ package dto;
 public class ItemDTO {
     private int iid;
     private String description;
-    private int qty_on_hand;
+    private double qty_on_hand;
     private double unit_price;
 
     public int getIid() {
@@ -22,11 +22,11 @@ public class ItemDTO {
         this.description = description;
     }
 
-    public int getQty_on_hand() {
+    public double getQty_on_hand() {
         return qty_on_hand;
     }
 
-    public void setQty_on_hand(int qty_on_hand) {
+    public void setQty_on_hand(double qty_on_hand) {
         this.qty_on_hand = qty_on_hand;
     }
